#150 mv likely rs layer stuff to new class
